[website] Bump Docusaurus to 3.1.1 and fix broken anchors (#3948)
The website build has started failing for an unknown reason, but there
was a warning about the Algolia integration and bumping Docusaurus seems
to fix the issue. Docusaurus 3.1.1 introduces checks for broken anchors
which is also fixed in this PR. One of the broken anchors was to a
section on FixedPoint, as FixedPoint was removed, this section was
deleted.

(cherry picked from commit 946392f5d1dd4aeafb4130f48f57d231db5b304f)

# Conflicts:
#	docs/src/cookbooks/cookbook.md