In BoringUtils.drive, don't bore into the final module for inputs.
This follows up on https://github.com/chipsalliance/chisel/pull/3960
to address the behavior at the final module for the sink being
driven. We already have special handling for boring out from the
original source, and we need the inverse here. When we reach the final
sink, if it is an input port, we should return it, rather than boring
into the module and connecting to it.