Flush staged commands before/after enter/exit regions.
Fix BoringUtils (see tests) placement of connection commands
so that they are added to the current command buffer.

Flush staged commands before changing regions to avoid needing
multiple "staging" buffer lists.
Alternatively the current list could be saved and restored.

Tests from #4108 are included and finished, with this change
two of them now pass (the ones encountered in practice, FWIW).
The third is included but marked "ignore" as future work.

This is a bit of a change regarding "staged" commands, as now
before the module is closed they are no longer "secret" since
they are added to the normal command buffers.
They are still staged (deferred) which appears to be useful
to ensure they appear after what's currently being constructed
(specifically, not before).