Make "deprecated public" binding APIs private (#4177)
* Make "deprecated public" binding APIs private

This is done by adding a new package private object "binding" within
which all bindings live. This required changing imports in lots of files
but it makes it much easier to not accidentally create a public binding
in the future.

* [cleanup] Delete @deprecated and private[chisel3] from Binding.scala

These are no longer needed now that all binding are packaged up in a
package private object.