Add Region to Chisel IR
Add a Region operation to Chisel IR.  This is a command which is a
container of other commands.  This is intended to provide an "insertion
point" that can be added to the command buffer which can then be
referenced later to add more commands.  This avoids the problem of the
command buffer being append only while _not_ going fully in the direction
of using a datastructure which allows arbitrary insertion (i.e., a linked
list).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>