Rework When IR to be Region-based (#4184)
Change the internal representation of `When` in Chisel to be region based
as opposed to token-based.  Concretely, change `WhenContext` to not push
token commands indicating when the region starts/stops, has an
alternative, etc.  Instead, use a single object that has an "if" and an
"else" region where commands can be pushed.

Co-authored-by: Asuna <SpriteOvO@gmail.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>