Encode ActualDirection as a Byte in Data
This does have a minor impact on public API around
ActualDirection.Bidirectional, but it's source compatible.

This saves 3 or 7 bytes (depending on heap size) per Data with
non-bidirectional ActualDirection (all bound hardware ends up with an
ActualDirection).

This saves ~20 bytes per Bidirectional Data.

Move _directionVar to be next to _specifiedDirection

This may help with alignment and reducing memory footprint.