Create layer blocks on demand
Change `chisel3.layer.block` to create parent layers on demand and only
error if the layer of the current layer block is not an ancestor of the
desired layer.  This is a relaxation of the previous requirement that the
current layer must be the parent of the desired layer.

The reason for this relaxation is to make future APIs easier to write.
E.g., there is a desire to put asserts under layers automatically.  Part
of such an API will need to automatically create layer blocks if the user
wants to put all assertions in a layer that is not a child of the root
layer.  Ergo, an API like what `block` was modified to be is necessary
either for `block` or in this new assertion code.  It seemed better to
make this available to all users and library authors.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>