[simulator] Factor out file movement function, NFC
Factor out a function to move a single file from a function that moves
files from a filelist.  This is done in preparation for later changes that
copy additional files _not_ in the filelist.  (This is part of broader
work to make directory control and layers work with Chisel simulators.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>