[simulator] Enable all layers for simulation
Change Chisel's simulator to enable all layers that exist in the circuit.
Do this by copying over the layer bind files and including them in
compilation.

This is a stopgap measure that just enables layers.  In the future, tests
will be able to specify what layers they want to enable.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>