[simulator] Add LayerControl infra (#4312)
Add some infrastructure for controlling which layers are enabled by
Chisel's simulator.  This is not currently exposed to the user, yet.  This
makes no changes to the existing behavior of all layers being enabled.

Add a function argument that can be used to choose (filter) which Verilog
files should be include in the build.  This is a _very_ low level API that
is being added to allow a user to enable bind convention layers when
writing a test.

Add the ability for the user of the EphemeralSimulator to control which
layers are enabled.

Add an option for controlling layers from ChiselTest-based tests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>