Add RemapLayer Chisel option, but do not use
Add a new annotation/option to Chisel, `RemapLayer`, which is intended to
be used to allow a Chisel compilation to change which layers are used by
an upstream project's libraries in a downstream project.  The motivation
here is to fix mismatches between upstream and downstream when they have
different layers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>