Add --remap-layers to ChiselStage
Add the ability for ChiselStage to globally remap a layer to another
layer.  This is intended to be used as a Chisel-level mechanism for a
downstream project to change the layers of all upstream projects.  This
may be necessary in order to produce compatible FIRRTL designs where the
upstream project and downstream project have different layers, but the
upstream project has put certain constructs (e.g., asserts) into
specific layers.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>