Add skipIfAlreadyInBlock arg to layer.block
Add an argument to `layer.block` which can be used to prevent a layer
block from being created if the current builder state indicates that we
are already inside a layer block.  This API is intended to be used for
library writers to provide a way to _maybe_ put an operation into a layer
block.  Specifically, this is intended for verification statements which
should go into a default layer block (e.g., `Verification.Assert`), but
should not do this if the Chisel user is already in another layer block.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>