Put verification statements in layers
Change assert, assume, cover, and stop APIs to emit their FIRRTL under one
of the newly defined layers in Chisel.  The layer where each of these is
placed follows the following scheme:

  - assert -> Verification.Assert
  - assume -> Verification.Assume
  - cover  -> Verification.Cover
  - stop   -> Verification.Stop

This is a change which may have implications for downstream projects and
Chisel users.  To remove the effects of this change, compile your design
with the following firtool option:

    -enable-layers=Verification,Verification::Assert,Verification::Assume,Verification::Cover

This will, inside the CIRCT compile, force these layers to always be
enabled.

If a user or downstream projects wants to put these statements into
alternative layers, they can simply use the APIs from within an
alternative `layer.block`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>