Add skipIfLayersEnabled param to layer.block
Add a second control parameter to `layer.block` that will cause no layer
to be created if the current module has any layers enabled.  This is
intended as a safe default for library writers that are putting operations
into layer blocks.  Certain verification operations do not work well with
other verification operations (e.g., force and assert) due to interactions
where layers may result in the creation of multiple always blocks.  This
option allows operations to be placed in the body of a testbench (the
likely module that has layers enabled) and avoid any of the interactions
of disjoint always blocks.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>