Add layer specification to LTL, verification ops (#4323)
Change assert, assume, cover, and stop APIs to emit their FIRRTL under one
of the newly defined layers in Chisel.  The layer where each of these is
placed follows the following scheme:

  - assert, AssertProperty -> Verification.Assert
  - assume, AssumeProperty -> Verification.Assume
  - cover, CoverProperty   -> Verification.Cover
  - stop                   -> Verification

This is a change which may have implications for downstream projects and
Chisel users.  To remove the effects of this change, compile your design
with the following firtool option:

    -enable-layers=Verification,Verification::Assert,Verification::Assume,Verification::Cover

This will, inside the CIRCT compile, force these layers to always be
enabled.

If a user or downstream projects wants to put these statements into
alternative layers, they can simply use the APIs from within an
alternative `layer.block`.  If a user wants to prevent these statements
from being placed in layers, they can enable a layer for the current
module.

Fix a bug in `DataEqualitySpec` which is incorrectly relying on the exact
ordering of a stop and assert to pass.  While this currently works, this
is problematic if these are legally reordered by extracting them (in
Chisel or a CIRCT compiler) into layers/bind modules.

Additionally, the `ProbeSpec` needed to be updated for this change because
it dependend on `force` and `assert` occurring with a defined procedural
order.  This commit may change the order of side-effecting statements.
This is sometimes problematic for statements which are inherently
procedural like `force`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>