Add layer.elideBlocks API
Add an API, `chisel3.layer.elideBlocks` that will cause any calls to
create a layer block inside its thunk to _not_ be created.  This is an
advanced API intended to be used in situations where a library has to
avoid additional layer block creation.

This is specifically being added because we have an API, backed by an
in-tree CIRCT transforms, which can create bound modules and does so
_after_ layers are converted to modules.  While this API is intended to be
replaced with layers, initially it needs to have a way to prevent layers
from being added to the design to prevent creation of a bind-under-bind in
the output.

I expect this API to be short lived...

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>