Make layer.addLayer API public
Make the `chisel3.layer.addLayer` API public.  This is useful if a user
wants to guarantee that a layer will show up in CHIRRTL even if no layer
block refers to this layer.

This solves a problem where downstream build flows want to expect that
certain layers will always exist.  However, without this API, there is no
way to do this other than to create dummy layer blocks.  This is slightly
cleaner without the empty layer block trick.  Additionally, there is no
real harm to having this API public.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>