Improve robustness of SimulatorSpec
Fix an issue with a test in the SimulatorSpec that was making assumptions
that `emitSystemVerilog` would _not_ produce multiple files.  Future proof
this test by only checking that the module's Verilog matches and ignore
the rest.

This change is made to allow for layers to be emitted by default.  This
will cause a bunch of (empty) bind files to show up and would break the
test.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>