Make '--module' support arguments
Dramatically increase the power of the '--module' argument to
'circt.stage.ChiselMain'.  This adds support for constructing modules with
parameters using reflection as demonstrated elsewhere [[1]].

This is long missing support that I intended to add when both the original
reflective construction was added [[2]] and turned on [[3]].  At the time,
I didn't know how to do reflective construction of arbitrary modules and
left that feature unfinished for the next 5 years.

This is done to move 'ChiselMain' more in the direction of being a command
line utility for running Chisel generators.  This is related to the
discussions of a hypothetical 'chisel-cli' that has been discussed on

This commit changes the API such that any class constructed this way must
take a trailing '()'.  Anything without this will be interpreted as a
string.  This is a change to how '--module' used to work where the
argument was assumed to be a class.

[1]: https://github.com/seldridge/reflective-builder
[2]: a423db5fa5a9fb106c1b0048b7dcf97fc83953b6
[3]: 4eff16bd76a461d76c52103b49889947ac1c9573

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>