Fix ActualDirection calculation from SpecifiedDirection (#4205)
Unspecified direction maps to Output while Flip maps to Input.
Previously, ActualDirection.fromSpecified would return Unspecified for
either Unspecified or Flipped input. This in turn resulted in Bundles
mixing Unspecified and Outputs as being "bidirectional" despite the fact
that they are actually unidirectional (or passive).

This change makes the direction calculation more consistent albeit at
the cost of marking the directions of unspecified Data as Output.