Wrap printfs in the Verification layerblock (#4506)
Change the emission of printfs to emit these under the `Verification`
layerblock.  This is done as part of replacing the `firtool`
extract-test-code feature with layers and the "advanced layer sink"
`firtool` feature.

The extract-test-code feature will extract asserts _and_ printfs
together. This is technically a deviation from this behavior.  However, it
seems better to not lump these in with asserts.  Nonetheless, this is a
backwards compatible change from the perspective of extract-test-code
because the `Verification` layer is always loaded if the `Assert` layer is
present.  (Prints will always show up if you turn on assertions.)

In the future, we probably want to add a `Debug` layer under the
`Verification` layer where these will live.  I would like to hold off on
making this change until extract-test-code is gone.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>