Add new InlineAllInstances trait
The InlineInstance trait can be attached to a module, to cause firtool to
inline instance of that module. The inline annotation is tricky to wield
because, if an inline module dedups with another module, we will inline all
instances of both modules.

To make InlineInstance easier to use, whenever we mark a module as inline, we
also mark it as "no dedup".

This of course does not work well when we have a module which we want to inline
AND dedup. So, this PR adds a new InlineAllInstances trait which allows a
module to be marked inline without blocking dedup.