fix: evaluate target lazily in Path
At the moment `toTarget` does not work properly with the D/I system,
seeing that we cannot get the instanceName of a module when evaluating
it, and even if we call `toTarget` in `atModuleBodyEnd` we will get the
wrong circuit name since the definition of the module's parent is still
in construction.

In order to construct `Path` properly, we need to defer the evaluation
of `target` to the IR generation time in the `isMember` constructor,
just as the other constructors already do.

Signed-off-by: unlsycn <unlsycn@unlsycn.com>