[test] Add Verilog test of inline emission, NFC (#4548)
* [test] Add Verilog test of inline emission, NFC

Add an integration test of SystemVerilog generation for inline layers.
This relies on new support in firtool 1.97.0 which adds support for
compiling this feature.  Previous versions of firtool could parse inline
layers, but error when compiling them.

* [test] Update btor2 emission test

There was a change in the btor2 emission upstream.  Update the test to
match the new emission.

I did not validate the correctness of this change.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>