Add layers to DesignAnnotation
Add a field to the `DesignAnnotation` that stores all layers that were
generated in a run of Chisel.  This is intended to be used by testing APIs
that need to know what layers exist in order to turn them all on or to
error if a user tries to enable a layer that does not exist.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>