[ChiselSim] Add inline layer control
Extend ChiselSim with the ability to control inline layers.  This requires
untangling some of the way that compilation settings are passed into a
`Simulator` because these options need to be added to after Chisel
elaboration.  E.g., to enable all inline layers, this requires knowing all
the layers that were created.  This didn't come up previously with extract
layers as all layer files could be blindly included for this case.

In susbequent commits, I will cleanup the way that extract layers work to
use the same code that is introduced here.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>