[ChiselSim] Better file include in LayerControl
Provide initial work that better handles the problem of needing to filter
files in order to enable specific layers.  This adds a new member
function, `shouldIncludeFile` which returns a `PartialFunction` that is
defined only if the file being examined is a layer file (as defined in the
FIRRTL ABI for extract layers) and returns true if that layer file should
be included in order to enable the requested layers.

In subsequent commits, this will be made load bearing.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>