[ChiselSim] Add non-ElaboratedModule LayerControl
Add functions for working with `LayerControl` member functions which do
not require an `ElaboratedModule`.  An `ElaboratedModule` is private and
basically impossible to work with unless you are already using
`ChiselSim`.  For projects that are extending `ChiselSim` (or have not
adopted it fully, yet), this makes `LayerControl` usable whereas without
this it is not.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>