[util] Add a withShadowLayer Queue
Add a method to the `Queue$` object that can be used to create a `Queue`
and include a second queue that contains some data which will be tracked
alongside the first `Queue`, but this will be kept in a layer.

This is added to help address a use case where design verification wants
to track some extra, sidecar data that should be added to some data
pipeline.  However, this sidecar data is expected to be removed from the
design.  This would normally be addressed with a generator-time parameter.
However, the generator-time parameter prevents removal of the sidecar data
unless the design is re-elaborated.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>