[util] Add a withShadowLayer Queue (#4589)
Add methods to `Queue` and `Queue$` that can be used for creating "shadow"
queues.  These are queues that are tracking other, auxiliiary data not in
the original queue and may possibly be instantiated elsewhere in the
design.

This is added in support of a design verification use case of building
"shadow pipelines" that track additional information about the execution
of some complicated pipeline (e.g., an out-of-order datapath) to expose
this information to tests or as part of debugging the pipeline.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>