adapt Scala3 syntax in ChiselEnum.scala and InstanceSpec.scala (#4598)
* rename enum to enumVal in ChiselEnum.scala

In Scala 3, 'enum' is a reserved keyword and cannot be used as a
variable name

* add parentheses around anonymous function parameters in InstanceSpec.scala

In Scala 3, the syntax for anonymous function parameters has changed to
require parentheses around the parameter list, even for single
parameters. This commit updates the code in InstanceSpec.scala to use
the new Scala 3 syntax by adding the required parentheses around the
anonymous function parameters.