[core] Add Placeholder API
Add an advanced Chisel API which allows for generation of hardware at a
remembered point in the design.  This API allows a user to create an
object of `class Placeholder` which, if used later via its `append`
method, will create hardware at the location where the `Placeholder`
object was created.

This is intended to be used to support certain generator patterns where
hardware must be created _before_ some other hardware, but only after the
other hardware has been built.  E.g., this can be used to support creating
and defining layer-colored probes after constructing a layer block.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>