[core] Add BlockReturnHandler type class
Add a type class, to be used for layer blocks, that can be used to
post-process a layer block's return value.  Two type class generators are
provided: one that is an identity function and another that returns a
wire of layer-colored probe type.

This latter type class generator has the effect of creating code like the
following without requiring a pre-declared layer:

    wire a: Probe<UInt<1>>
    layerblock A:
      wire b: UInt<1>
      define a = probe(b)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>