[core] Allow return values from layer.block
Change the `layer.block` API to return values.  If the value is a `Data`,
then create a wire of layer-colored probe type immediately before the
layer block.  Inside the layer block, define this value with a probe of
the return.  If the value is a non-`Data`, then return it directly.

Previously, this API would always return `Unit`.

h/t @jackkoenig for design of the original type class approach.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>