[core] Add BlockReturnHandler type class
Add a type class that can be used to do special handling of values
returned from layer blocks.  This effectively causes values of type `Data`
to become a wire of layer-colored probe type driven by the value returned.
For any other type, this will return `Unit`.

This will be used in a later commit to change the behavior of layer blocks
to work in the above way.

Co-authored-by: Jack Koenig <koenig@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>