[core] Enable returning values from layer blocks
Add support for a limited form of returning values from layer blocks.
With this patch, when a `Data` is returned from a layer block, this will
now cause the creation of a wire of layer-colored probe type that is
driven by a probe of the return value.  There is additional handling to
directly return a probe type as well (which enables returning from an
inner layer block up through an outer layer block).  If a non-`Data` is
returned, then this will be converted to a `Unit`.  This latter behavior
is the current behavior.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>