[svsim] Use 2-state variable for ports
Fix an apparent additional problem with svsim and VCS xprop.  For reasons
that I am unable to comprehend, this change gets xprop simulations to
pass.  Without them, I am continuing to see spurious clock toggles causing
assertions to fire.

This is a follow-on to a77eee6.  This is a bit of a game of whack-a-mole
at this point...

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>