[chiselsim] Add Default (non-Ephemeral) Simulator
Add a new simulator, DefaultSimulator, to Chiselsim.  This fills a major,
long-standing hole in this API by adding a simulator that does not delete
all the files that it creates after a test runs.

This uses a type class, `HasTestingDirectory`, to allow for control over
where the directory is created.  A default implementation is provided
which uses a nanosecond resolution timestamp to differentiate the tests.
E.g., example output from running the new test included in this commit
multiple times is:

    test_run_dir/
    └── DefaultSimulator
        ├── 2025-02-05T16-58-02.175175
        ├── 2025-02-05T16-58-11.941263
        └── 2025-02-05T16-58-17.721776

Note: a Scalatest-based type class implementation is intentionally _not_
provided as part of this PR.  This will be provided in a follow-on in such
a way that Chiselsim does not inherit a dependency on Scalatest.  Ideally,
users of Scalatest will use the future Scalatest-based type class
implementation to get a subdirectory which is not a timestamp, but derived
from the name of that part of the test.  I.e., a directory like:

    test_run_dir/
    └── DefaultSimulator
        ├── writes_files_to_a_permanent_directory_on_disk
        └── doesnt_delete_the_root_directory

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>