[Chiselsim] Add WithTestingDirectory (#4669)
Add a trait, `WithTestingDirectory`, that will automatically create a test
directory structure based on Scalatest information.  This provides a type
class implementation of `HasTestingDirectory`, but with better
Scalatest-specific directories as opposed to generic timestamp-based ones.

This necessarily introduces a provided ("compile only" to Mill) dependency
on Scalatest for Chisel.  This prevents making all of Chisel pull in
Scalatest and relies on the user to load it.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>