[chiselsim] General cleanup (#4674)
Remove an unused trait that was intending to abstract over multiple runs
of a simulation using different simulators.  Instead, this can be handled
directly in Scalatest (via an argument) or via the build system.

Remove an abstract method that overrides an existing abstract method.

Collapse single-use traits, methods, and classes into their usage sites.

Fix an incorrect filename in a test.

Fix a bug where the DefaultSimulator would not properly error if an expect
failed.  Fix this and add a test that it does fail.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>