[chiselsim] Add HasSimulator type class
Add a new type class, `HasSimulator`, which provides a method to get a
simulator.  Provide two implementations: one for Verilator and one for
VCS.  Additionally, provide a low-priority default that maps to Verilator.

This is working towards refactoring Chiselsim to improve the
customizability of its APIs while not dramatically changing how a user
interacts with it.  Additionally, this is trying to work towards providing
Scalatest-specific implementations of `HasSimulator` which allow the
simulator to be chosen on the command line.  Much of the code for this is
currently sequestered in either ChiselSpec (part of the tests which are
not published) or in internal code.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>