[chiselsim] Add ChiselSim trait
Add two new traits:

  1. chisel3.simulator.ChiselSim
  2. chisel3.simulator.scalatest.ChiselSim

They provide the same set of APIs (`PeekPokeAPI` and `SimulatorAPI`), but
the latter is additionally specialized for better integration with
Scalatest (and also requires integration with Scalatest via a self type).

These are intended to be the main entrypoints for working with the
platonic ideal of "ChiselSim" and are not supposed to be anything other
than a collection of APIs.  (Please don't add anything more to these and
instead factor additional APIs into additions to their constituent
traits.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>