[chiselsim] Improve error behavior
Fix longstanding problems in Chiselsim where the result of a failing
simulation (e.g., an assertion firing) would result in an obtuse
`UnexpectedEndOfMessages` assertion.  This then forces any Chiselsim user
to catch this and then figure out what went wrong.  See the unpublished
ChiselSpec for an example.

Fix this with an abstract method for the Backend that forces it to
describe (via a RegEx) how to find interesting parts of the log file which
indicate certain failures (the aforementioned assertion failures).  This
then allows for the default simulate method to substitute a _better_
assertion where possible.

This is not as good as a more advanced alternative which actually hooks
into the simulation to do something when an assertion fires.  However,
this is a larger effort that requires changing the Verilog that is emitted
for CIRCT-compiled FIRRTL (or defining an ABI for assertions).  We
definitely want to move in this direction, though.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>